+++ products/views.py
@@ -142,6 +142,10 @@ def delete_product(request, product_id):
         return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.info(request, 'The painting has been deleted!')
-    return redirect(reverse('products'))
+    if request.method == 'POST':
+        product.delete()
+        messages.info(request, 'The painting has been deleted!')
+        return redirect('products')
+    context = {'product': product}
+    return render(request, 'products/delete_product.html', context)
+
